requirements.txt
Pillow==9.2.0
streamlit==1.12.0
streamlit-lottie==0.0.3
requests==2.28.1
